rename in mapboxsimple
