+++ plasmic-library/components.ts
@@ -150,4 +150,4 @@ export { default as PageLoaderMeta } from "./others/PageLoader/PageLoader.meta";
 export { default as SmartLoader } from "./others/SmartLoader/SmartLoader";
 export { default as SmartLoaderMeta } from "./others/SmartLoader/SmartLoader.meta";
 
-export { default as MapBox } from "./MapBox/MapBox.meta";
+export { default as MapBoxSimple } from "./MapBoxSimple/MapBoxSimple.meta";
